Move abstract models to datahubhel/base_models.py
Move the abstract models to datahubhel/base_models.py file. Since
the utils/models.py is no longer needed, delete it.
