+++ backend/datahubhel/base_models.py
@@ -1,0 +1,25 @@
+import uuid
+
+from django.conf import settings
+from django.db import models
+from django_extensions.db.models import TimeStampedModel
+
+
+class TimestampedUUIDModel(TimeStampedModel):
+    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+
+    class Meta:
+        abstract = True
+
+
+class EntityBase(models.Model):
+    owner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, null=True, blank=True)
+    sts_id = models.CharField(max_length=255, unique=True)
+    name = models.CharField(max_length=255, null=True, blank=True)
+    description = models.TextField(null=True, blank=True)
+
+    class Meta:
+        abstract = True
+
+    def __str__(self):
+        return self.name
