+++ backend/dhh_auth/models.py
@@ -10,7 +10,7 @@ from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
 from dhh_auth.utils import get_perm_obj
-from utils.models import TimestampedUUIDModel
+from datahubhel.base_models import TimestampedUUIDModel
 
 
 class ClientPermission(TimestampedUUIDModel):
