+++ backend/gatekeeper/models.py
@@ -1,18 +1,7 @@
 from django.conf import settings
 from django.db import models
 
-
-class EntityBase(models.Model):
-    owner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, null=True, blank=True)
-    sts_id = models.CharField(max_length=255, unique=True)
-    name = models.CharField(max_length=255, null=True, blank=True)
-    description = models.TextField(null=True, blank=True)
-
-    class Meta:
-        abstract = True
-
-    def __str__(self):
-        return self.name
+from datahubhel.base_models import EntityBase
 
 
 class Thing(EntityBase):
